requirements-gpu.txt
tensorflow-gpu==2.5.0
opencv-python~=4.5.2
lxml
tqdm
absl-py
matplotlib
easydict
pillow
scipy